repartition des pages du dashboard en dossier
