+++ src/components/dashboard/help/help.js
