+++ src/components/dashboard/transactions/transactions.js
@@ -1,8 +1,11 @@
 import React, { Component } from 'react';
 import  {GridRow, Container, Grid, Image } from 'semantic-ui-react';
-import Layout from './withSideBarLayout'
+import Layout from '../withSideBarLayout'
 
-import Calendar from '../../../assets/Calendar.svg'
+import Calendar from '@assets/Calendar.svg'
+
+
+
 
 export const FilterButton = (props) => {
     return (
