+++ styles/pages/accounts.css
@@ -99,8 +99,7 @@
 }
 .link__paypal {
     margin-top: 1.5rem;
-    padding: 1.5rem !important;
-    font-size: 1.5rem !important;
+    display: inline-block;
 }
 
 /*************  quick overview  ********/
@@ -111,7 +110,7 @@
     width: 100% !important;
     max-width: 45rem !important;
     background-color: var(--white-color) !important;
-    padding: 2rem !important;
+    padding: 2.5rem 3rem !important;
     margin: 1rem 0px !important;
     /* border-radius: 4px; */
     /* border: 1px solid #bbb; */
@@ -119,8 +118,8 @@
 }
 
 .bg_title {
-    font-size: 2.5rem;
-    font-weight: 500;
+    font-size: 2rem;
+    font-weight: 600;
     display: inline-block;
     margin-bottom: 1rem;
     color: var(--black-color);
@@ -143,7 +142,7 @@
 }
 
 .sum {
-    font-size: 3rem;
+    font-size: 3.5rem;
     font-weight: 800;
 }
 
@@ -182,3 +181,41 @@
     bottom: 1rem;
     right: 1rem;
 }
+
+
+/********** New Items *******/
+/* .newItems__link {
+    display: block;
+    padding: 1rem;
+    background-color: beige;
+}
+
+.newItems__parent {
+    display: flex !important;
+}
+
+.newItems__profil {
+
+}
+
+.newItems__tof {
+    background-color: aqua;
+    height: 5rem;
+    width: 6rem;
+}
+
+.newItems__infos {
+    flex: 1 0;
+    padding-left: 1rem;
+    display: flex;  
+}
+
+.newItems__seller {
+    width: 25%;
+}
+.newItems__prices {
+    width: 25%;
+}
+.newItems__stats {
+    width: 45%;
+} */
